requirements.txt
django==2.1.8
enum-compat==0.0.2
eventlet==0.21.0
greenlet==0.4.12
python-engineio
python-socketio
pytz==2018.7
six==1.10.0
